+++ glean-core/android/src/test/java/mozilla/telemetry/glean/GleanTest.kt
@@ -176,6 +176,12 @@ class GleanTest {
         assertFalse(Glean.testIsExperimentActive("experiment_preinit_disabled"))
     }
 
+    @Test
+    fun `test experimentation id recording`() {
+        resetGlean(config = Configuration(experimentationId = "alpha-beta-gamma-delta"))
+        assertEquals("alpha-beta-gamma-delta", Glean.testGetExperimentationId())
+    }
+
     // Suppressing our own deprecation before we move over to the new event recording API.
     @Test
     @Suppress("ComplexMethod", "LongMethod", "NestedBlockDepth", "DEPRECATION")
@@ -472,6 +478,7 @@ class GleanTest {
             name = pingName,
             includeClientId = true,
             sendIfEmpty = false,
+            preciseTimestamps = true,
             reasonCodes = listOf(),
         )
         val stringMetric = StringMetricType(
@@ -837,6 +844,7 @@ class GleanTest {
             name = pingName,
             includeClientId = true,
             sendIfEmpty = false,
+            preciseTimestamps = true,
             reasonCodes = listOf(),
         )
         val stringMetric = StringMetricType(