+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v54.0.0...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v55.0.0...main)
+
+# v55.0.0 (2023-10-23)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v54.0.0...v55.0.0)
+
+* General
+  * BREAKING CHANGE: Adding `0` to a `counter` or `labeled_counter` metric will be silently ignored instead of raising an `invalid_value` error ([bug 1762859](https://bugzilla.mozilla.org/show_bug.cgi?id=1762859))
+  * Trigger the uploader thread after scanning the pending pings directory ([bug 1847950](https://bugzilla.mozilla.org/show_bug.cgi?id=1847950))
+  * Extend start/stop time of a ping to millisecond precision. Custom pings can opt-out using `precise_timestamps: false` ([#2456](https://github.com/mozilla/glean/pull/2456))
+  * Update `glean_parser` to v10.0.0. Disallow `unit` field for anything but quantity, disallows `ping` lifetime metrics on the events ping, allows to configure precise timestamps in pings ([release notes](https://github.com/mozilla/glean_parser/releases/tag/v10.0.0))
 
 # v54.0.0 (2023-09-12)
 
@@ -8,6 +18,7 @@
 
 * General
   * Experimental: Add configuration to add a wall clock timestamp to all events ([#2513](https://github.com/mozilla/glean/issues/2513))
+  * Add an API to set an Experimentation ID that will be annotated to all pings ([Bug 1848201](https://bugzilla.mozilla.org/show_bug.cgi?id=1848201))
 * Python
   * Switched the build system to maturin. This should not have any effect on consumers. ([#2345](https://github.com/mozilla/glean/pull/2345))
   * BREAKING CHANGE: Dropped support for Python 3.6 ([#2345](https://github.com/mozilla/glean/pull/2345))